Arm backend: Create more general GetDecompositionPass
* Add GetDecompositiobPass for passes that use get_decomposition()
